youtub/requirements.txt
fastapi==0.116.1
google-genai==1.34.0
pydantic==2.11.7
uvicorn==0.35.0
